prpatt split into geo_data_utils and pattern_logic_lib, refactoring and improved test coverage
